allow hierarchical inserts into LDAP from OVSDb
with this changeset, inserts into LDAP can be made in a hierarchy. For
example, Logical_Router is the parent of Logical_Router_Port, the insert
query from OVSDb Mgmt Protocol should be reorganized in a hierarchical
form. The resulting JSON file will include "parent" as an entry if it
should be entered in hierarchy.

This changeset does not support multi-level hiearchy inserts. For
changing the resulting JSON entry for LDAP, please modify
transact_aggregate to fit your needs.